clean up references to posix
