+++ std/c/linux.zig
@@ -2,6 +2,7 @@ const std = @import("../std.zig");
 use std.c;
 
 pub extern "c" fn getrandom(buf_ptr: [*]u8, buf_len: usize, flags: c_uint) c_int;
+pub extern "c" fn sched_getaffinity(pid: c_int, size: usize, set: *cpu_set_t) c_int;
 extern "c" fn __errno_location() *c_int;
 pub const _errno = __errno_location;
