+++ std/crypto.zig
@@ -33,7 +33,7 @@ pub const Poly1305 = @import("crypto/poly1305.zig").Poly1305;
 pub const X25519 = @import("crypto/x25519.zig").X25519;
 
 const std = @import("std.zig");
-pub const randomBytes = std.posix.getrandom;
+pub const randomBytes = std.os.getrandom;
 
 test "crypto" {
     _ = @import("crypto/blake2.zig");
