+++ std/os/linux/x86_64.zig
@@ -1,4 +1,4 @@
-use @import("posix/x86_64.zig");
+use @import("../bits.zig");
 
 pub fn syscall0(number: usize) usize {
     return asm volatile ("syscall"
